src/experiments/agentify_tau_bench/pyproject.toml
[project]
name = "agentify-tau-bench"
version = "0.1.0"
description = "Agentify Tau-Bench using A2A"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "dotenv>=0.9.9",
    "tau2",
    "typer<0.16",
    "uvicorn>=0.37.0",
    "gymnasium>=0.31.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
]

[tool.uv.sources]
tau2 = { path = "../../..", editable = true }

[project.scripts]
tau-bench-agent = "agentify_tau_bench.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentify_tau_bench"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["E501", "F401", "F541"]

[tool.ruff]
line-length = 88

[tool.mypy]
strict = true
python_version = "3.13"
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
